refactor layout folder
