--- src/app/organizations/[organization]/members/layout.tsx
@@ -1,5 +1,0 @@
-import OrgSidebar from '../../../../layouts/OrgSidebar';
-
-export default function MembersLayout({ children }: { children: React.ReactNode }): JSX.Element {
-  return <OrgSidebar>{children}</OrgSidebar>;
-}
