--- src/app/organizations/[organization]/projects/layout.tsx
@@ -1,5 +1,0 @@
-import OrgSidebar from '../../../../layouts/OrgSidebar';
-
-export default function ProjectsLayout({ children }: { children: React.ReactNode }): JSX.Element {
-  return <OrgSidebar>{children}</OrgSidebar>;
-}
