--- src/app/organizations/[organization]/settings/layout.tsx
@@ -1,5 +1,0 @@
-import OrgSidebar from '../../../../layouts/OrgSidebar';
-
-export default function SettingsLayout({ children }: { children: React.ReactNode }): JSX.Element {
-  return <OrgSidebar>{children}</OrgSidebar>;
-}
